// WCI/src/app/models/manage-integration-item.ts
export class ManageIntegrationItem {
    Address: string;
    Channel: string;
    CharacterEncoding: any;
    Connected: boolean;
    Id: any = 0;
    IsServer: boolean;
    Name: string = '';
    NavigationUrl: string;
    Port: number;
    Status:any;
    System:any;
    Type:any;
    isExpanded: boolean = true;
    Version: number;
    TypeDescription: string = 'Central Link';
    isFieldDisabled: boolean;
}